+++ src/pyneng_cli_course/exceptions.py
@@ -1,0 +1,6 @@
+class PynengError(Exception):
+    """
+    Ошибка в использовании/работе скрипта pyneng
+    """
+
+
